pyproject.toml
[project]
name = "bmad-automate"
version = "0.2.0"
description = "Automated BMAD Workflow Orchestrator — powered by GitHub Copilot CLI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Robert Guss", email = "robertguss@gmail.com" },
]
keywords = ["bmad", "automation", "workflow", "copilot", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.9.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "httpx>=0.24",
]

[project.urls]
Homepage = "https://github.com/robertguss/bmad-automate"
Repository = "https://github.com/robertguss/bmad-automate"
Issues = "https://github.com/robertguss/bmad-automate/issues"

[project.scripts]
bmad-automate = "bmad_automate.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bmad_automate"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "ANN", "B", "C4", "SIM"]
ignore = ["ANN401"]
